+++ handlers/handlers_manager.go
@@ -8,5 +8,6 @@ func LoadHandlers() {
 	http.HandleFunc("/", homeHandler)
 	http.HandleFunc("/echo", echoRequestHandler)
 	http.HandleFunc("/createEchoRequest", createEchoRequestHandler)
+	http.HandleFunc("/getAPIResponse", getAPIResponse)
 	http.Handle("/static/images/", http.StripPrefix("/static/images/", http.FileServer(http.Dir("static/images/"))))
 }
