+++ static/template/api_response.html
@@ -1,0 +1,20 @@
+{{define "body"}}
+<div class="mx-5">
+    <form method="POST" action="/getAPIResponse">
+        <div class="form-group mt-5">
+            <label for="formControlSelect1">HTTP Method</label>
+            <select class="form-control" id="http-method" name="resp-type">
+                <option>GET</option>
+                <option>POST</option>
+            </select>
+        </div>
+        <div class="input-group mb-3">
+            <span class="input-group-text" id="inputGroup-sizing-default">URL</span>
+            <input type="text" class="form-control" aria-label="Sizing example input"
+                aria-describedby="inputGroup-sizing-default">
+        </div>
+        <button type="submit" class="btn btn-primary">Submit</button>
+        <button onclick="history.back()" type="button" class="btn btn-primary ml-4">Back</button>
+    </form>
+</div>
+{{end}}
