+++ static/template/default_template.html
@@ -3,14 +3,14 @@
 <html lang="en">
 
 <head>
-    <nav class="navbar navbar-light bg-light">
+    <nav class="navbar">
         <div class="container-fluid">
-            <a class="navbar-brand" href="#">
+            <a class="navbar-brand " href="/">
                 <img src="static/images/robot_icon.png" alt="Robot Logo" style="width: 75px; height: 75px"
                     class="d-inline-block align-text-center">
-                EchoBot
             </a>
         </div>
+
     </nav>
     <!-- Required meta tags -->
     <meta charset="utf-8">
