+++ static/template/echo_request.html
@@ -24,6 +24,7 @@
             <textarea class="form-control" id="responseBodyRequested" name="responseBodyRequested" rows="10"></textarea>
         </div>
         <button type="submit" class="btn btn-primary">Submit</button>
+        <button onclick="history.back()" type="button" class="btn btn-primary ml-4">Back</button>
     </form>
 </div>
 {{end}}
