+++ static/template/home.html
@@ -1,0 +1,27 @@
+{{define "body"}}
+<div class="container">
+    <div class="card text-center">
+        <div class="card-header">
+            EchoBot Activated!
+        </div>
+        <div class="row">
+            <div class="col">
+                <div class="card-body">
+                    <h5 class="card-title">Create An Echo Request</h5>
+                    <p class="card-text"></p>
+                    <a href="/createEchoRequest" class="btn btn-primary">Lets Go</a>
+                </div>
+            </div>
+            <div class="col">
+                <div class="card-body">
+                    <h5 class="card-title">Get An API Response</h5>
+                    <p class="card-text"></p>
+                    <a href="callAPI" class="btn btn-primary">Lets Go</a>
+                </div>
+            </div>
+        </div>
+        <div class="card-footer text-muted">
+        </div>
+    </div>
+</div>
+{{end}}
